requirements.txt
pillow
scipy
torch
torchvision
tqdm>=4.41.0,<5
flask
werkzeug